+++ Unity/Assets/Scripts/Main Game/Scriptable Objects/Bullets/ShotType.cs
@@ -8,7 +8,7 @@ namespace KH
         public GameObject bulletPrefab;
         public Sprite sprite;
         public float speed;
-        public float damage;
+        public int damage;
         public Vector2 spawnOffset1;
         public Vector2 spawnOffset2;
     }
